+++ src/derivkit/__init__.py
@@ -3,7 +3,7 @@
 from .adaptive_fit import AdaptiveFitDerivative
 from .finite_difference import FiniteDifferenceDerivative
 from .forecast_kit import ForecastKit
-from .kit import DerivativeKit
+from .derivative_kit import DerivativeKit
 from .plotutils.plot_helpers import PlotHelpers
 from .plotutils.plot_kit import PlotKit
 from .utils import (
@@ -28,4 +28,5 @@ __all__ = [
     "is_symmetric_grid",
     "generate_test_function",
     "ForecastKit"
+    "LikelihoodExpansion"
 ]
