+++ src/derivkit/forecast_kit.py
@@ -1,300 +1,49 @@
-"""Provides tools for facilitating experimental forecasts.
+# src/derivkit/forecast_kit.py
+"""Provides the ForecastKit class.
 
-The user must specify the observables, fiducial values and covariance matrix
-at which the derivative should be evaluated. Derivatives of the first order
-are Fisher derivatives. Derivatives of second order are evaluated using the
-derivative approximation for likelihoods (DALI) technique as described in
-https://doi.org/10.1103/PhysRevD.107.103506.
-
-More details about available options can be found in the documentation of
-the methods.
+A light wrapper around :class:`LikelihoodExpansion` that exposes a simple
+API for Fisher and DALI tensors.
 
 Typical usage example:
 
->>>    forecaster = ForecastKit(function=model, theta0=theta0, cov=cov)
->>>    fisher = forecaster.fisher()
->>>   g_tensor, h_tensor = forecaster.dali()
+>>> forecaster = ForecastKit(function=model, theta0=theta0, cov=cov)
+>>> F = forecaster.fisher()
+>>> G, H = forecaster.dali()
 """
 
-from copy import deepcopy
-
+from typing import Callable, Sequence
 import numpy as np
 
-from derivkit.kit import DerivativeKit
+from derivkit.likelihood_expansion import LikelihoodExpansion
 
 
 class ForecastKit:
-    """Provides tools for facilitating experimental forecasts.
+    """Provides access to Fisher and DALI likelihood-expansion tensors.
 
     Attributes:
-         function (callable): The scalar or vector-valued function to
-             differentiate. It should accept a list or array of parameter
-             values as input and return either a scalar or a
-             :class:`np.ndarray` of observable values.
-         theta0 (class:`np.ndarray`): The point(s) at which the
-             derivative is evaluated. A 1D array or list of parameter values
-             matching the expected input of the function.
-         cov (class:`np.ndarray`): The covariance matrix of
-             the observables. Should be a square matrix with shape
-             (n_observables, n_observables), where n_observables is the
-             number of observables returned by the function.
-         n_parameters (int): The number of elements of `theta0`.
-         n_observables (int): The number of cosmic observables. Determined
-             from the dimension of `cov`.
+        fisher: Method returning the Fisher information matrix (shape (P, P)).
+        dali: Method returning the doublet-DALI tensors G and H (shapes (P,P,P) and (P,P,P,P)).
     """
 
-    def __init__(self, function, theta0, cov):
-        """Initialises the class.
+    def __init__(
+        self,
+        function: Callable[[Sequence[float] | np.ndarray], np.ndarray],
+        theta0: Sequence[float] | np.ndarray,
+        cov: np.ndarray,
+    ):
+        """Initialises the forecaster with model, fiducials, and covariance.
 
         Args:
-            function (callable): The scalar or vector-valued function to
-                differentiate. It should accept a list or array of parameter
-                values as input and return either a scalar or a
-                :class:`np.ndarray` of observable values.
-            theta0 (class:`np.ndarray`): The points at which the
-                derivative is evaluated. A 1D array or list of parameter values
-                matching the expected input of the function.
-            cov (class:`np.ndarray`): The covariance matrix of
-                the observables. Should be a square matrix with shape
-                (n_observables, n_observables), where n_observables is the
-                number of observables returned by the function.
-
-        Raises:
-            ValueError: raised if cov is not a square numpy array.
+            function: Model mapping parameters -> observables (1D array-like in, 1D array out).
+            theta0: Fiducial parameter values (shape (P,)).
+            cov: Observables covariance (shape (N, N)).
         """
-        self.function = function
-        self.theta0 = np.atleast_1d(theta0)
-        if not cov.ndim < 3:
-            raise ValueError(
-                        "cov must be at most two-dimensional but "
-                        f"dimensions is {cov.ndim}."
-            )
-        if cov.ndim == 2 and not cov.shape[0] == cov.shape[1]:
-            raise ValueError("cov must be a square numpy array.")
-        self.cov = cov
-        self.n_parameters = len(self.theta0)
-        self.n_observables = len(cov)
+        self._lx = LikelihoodExpansion(function, theta0, cov)
 
-    def fisher(self, *, n_workers: int = 1):
-        """Return the Fisher information matrix (P,P)."""
-        return self.get_forecast_tensors(forecast_order=1, n_workers=n_workers)
+    def fisher(self, *, n_workers: int = 1) -> np.ndarray:
+        """Return the Fisher information matrix with shape (P, P)."""
+        return self._lx.fisher(n_workers=n_workers)
 
     def dali(self, *, n_workers: int = 1):
-        """Return the doublet-DALI tensors (G,H) with shapes (P,P,P) and (P,P,P,P)."""
-        return self.get_forecast_tensors(forecast_order=2, n_workers=n_workers)
-
-    def get_forecast_tensors(self, forecast_order=1, n_workers=1):
-        """Returns a set of tensors according to the requested order of the forecast.
-
-        Args:
-            forecast_order (int): The requested order D of the forecast:
-
-                    - D = 1 returns a Fisher matrix.
-                    - D = 2 returns the 3-d and 4-d tensors required for the
-                      doublet-DALI approximation.
-                    - D = 3 would be the triplet-DALI approximation.
-
-                Currently only D = 1, 2 are supported.
-            n_workers (int, optional): Number of worker to use in multiprocessing.
-                Default is 1 (no multiprocessing).
-
-        Returns:
-            :class:`np.ndarray`: A list of numpy arrays:
-
-                    - D = 1 returns a square matrix of size n_parameters, where
-                      n_parameters is the number of parameters included in the
-                      forecast.
-                    - D = 2 returns one array of shapes
-                      (n_parameters, n_parameters, n_parameters) and one array
-                      of shape (n_parameters, n_parameters, n_parameters, n_parameters),
-                      where n_parameters is the number of parameters included
-                      in the forecast.
-
-        Raises:
-            ValueError: A ValueError occurs when a forecase order greater than
-                2 is requested.
-            Exception: An exception occurs if the covariance matrix cannot
-                be inverted.
-            RunTimeError: A RunTimeError occurs if the ValueError was not
-                raised when calling this function.
-        """
-        if forecast_order not in [1, 2]:
-            raise ValueError(
-                "Only Fisher (order 1) and doublet-DALI (order 2) forecasts are currently supported."
-            )
-
-        # Invert the covariance matrix
-        try:
-            invcov = np.linalg.inv(self.cov)
-        except np.linalg.LinAlgError:
-            print("Standard inversion failed. Trying pseudoinverse.")
-            try:
-                invcov = np.linalg.pinv(
-                    self.cov
-                )
-            except Exception as e:
-                print(f"Pseudoinverse also failed: {e}")
-                invcov = np.full(
-                    (self.n_observables, self.n_observables), np.nan
-                )
-
-        if forecast_order == 1:
-            # Compute Fisher matrix
-            dfp1 = self._get_derivatives(
-                derivative_order=1, n_workers=n_workers
-            )
-            # F_ab = sum(i,j) df_i/dp_a * InvCov_ij * df_j/dp_b
-            fisher_ab = np.einsum(
-                "ai,ij,bj->ab", dfp1, invcov, dfp1
-            )
-            return fisher_ab
-
-        elif forecast_order == 2:
-            # Compute doublet-DALI tensors
-            dfp1 = self._get_derivatives(
-                derivative_order=1, n_workers=n_workers
-            )
-            dfp2 = self._get_derivatives(
-                derivative_order=2, n_workers=n_workers
-            )
-            # G_abc = sum(i,j) df_i/(dp_a dp_b )* InvCov_ij * df_j/dp_c
-            g_abc = np.einsum(
-                "abi,ij,cj->abc", dfp2, invcov, dfp1
-            )
-            # H_abcd = sum(i,j) df_i/(dp_a dp_b) * InvCov_ij * df_j/(dp_c dp_d)
-            h_abcd = np.einsum(
-                "abi,ij,cdj->abcd", dfp2, invcov, dfp2
-            )
-            return g_abc, h_abcd
-
-        raise RuntimeError("Unreachable code reached in get_forecast_tensors.")
-
-    def _get_derivatives(self, derivative_order, n_workers=1):
-        """Returns derivatives of the observables of the requested order.
-
-        Args:
-            derivative_order (int): The requested order d of the derivatives:
-
-                - d = 1 returns first-order derivatives.
-                - d = 2 returns second-order derivatives.
-
-                Currently only d = 1, 2 are supported.
-
-            n_workers (int, optional): Number of worker to use in
-                multiprocessing. Default is 1 (no multiprocessing).
-
-        Returns:
-            :class:`np.ndarray`: An array of derivative values:
-
-                - d = 1 returns an array with shape
-                  (`n_parameters`, `n_observables`) containing
-                  first-order derivatives.
-                - d = 2 returns an array with shape
-                  `n_parameters`, `n_parameters`, `n_observables`)
-                  containing second-order derivatives.
-
-        Raises:
-            ValueError: An error occurred if a derivative was requested of
-                higher order than 2.
-            RuntimeError: An error occurred if a ValueError was not raised
-                after calling the function.
-        """
-        if derivative_order not in [1, 2]:
-            raise ValueError(
-                "Only first- and second-order derivatives are currently supported."
-            )
-
-        if derivative_order == 1:
-            # Get the first-order derivatives
-            first_order_derivatives = np.zeros(
-                (self.n_parameters, self.n_observables)
-            )
-            for m in range(self.n_parameters):
-                # 1 parameter to differentiate, and n_parameters-1 parameters to hold fixed
-                theta0_x = deepcopy(self.theta0)
-                function_to_diff = self._get_partial_function(
-                    self.function, m, theta0_x
-                )
-                kit = DerivativeKit(function_to_diff, self.theta0[m])
-                first_order_derivatives[m] = kit.adaptive.compute(
-                    derivative_order=1, n_workers=n_workers
-                )
-            return first_order_derivatives
-
-        elif derivative_order == 2:
-            # Get the second-order derivatives
-            second_order_derivatives = np.zeros(
-                (self.n_parameters, self.n_parameters, self.n_observables)
-            )
-            for m1 in range(self.n_parameters):
-                for m2 in range(self.n_parameters):
-                    if m1 == m2:
-                        # 1 parameter to differentiate twice, and n_parameters-1 parameters to hold fixed
-                        theta0_x = deepcopy(self.theta0)
-                        function_to_diff1 = self._get_partial_function(
-                            self.function, m1, theta0_x
-                        )
-                        kit1 = DerivativeKit(
-                            function_to_diff1, self.theta0[m1]
-                        )
-                        second_order_derivatives[m1][m2] = (
-                            kit1.adaptive.compute(
-                                derivative_order=2, n_workers=n_workers
-                            )
-                        )
-
-                    else:
-                        # 2 parameters to differentiate once, with other parameters held fixed
-                        def function_to_diff2(y):
-                            theta0_y = deepcopy(self.theta0)
-                            theta0_y[m2] = y
-                            function_to_diff1 = self._get_partial_function(
-                                self.function, m1, theta0_y
-                            )
-                            kit1 = DerivativeKit(
-                                function_to_diff1, self.theta0[m1]
-                            )
-                            return kit1.adaptive.compute(derivative_order=1)
-
-                        kit2 = DerivativeKit(
-                            function_to_diff2, self.theta0[m2]
-                        )
-                        second_order_derivatives[m1][m2] = (
-                            kit2.adaptive.compute(
-                                derivative_order=1, n_workers=n_workers
-                            )
-                        )
-
-            return second_order_derivatives
-
-        raise RuntimeError("Unreachable code reached in get_forecast_tensors.")
-
-    def _get_partial_function(
-        self, full_function, variable_index, fixed_values
-    ):
-        """Returns a single-variable version of a multivariate function.
-
-        A single parameter must be specified by index. AAll others parameters
-        are held fixed.
-
-        Args:
-            full_function (callable): A function that takes a list of
-                n_parameters parameters and returns a vector of n_observables
-                observables.
-            variable_index (int): The index of the parameter to treat as the
-                variable.
-            fixed_values (list or np.ndarray): The list of parameter values to
-                use as fixed inputs for all parameters except the one being
-                varied.
-
-        Returns:
-            callable: A function of a single variable, suitable for use in
-                differentiation.
-        """
-
-        def partial_function(x):
-            params = deepcopy(fixed_values)
-            params[variable_index] = x
-            return np.atleast_1d(full_function(params))
-
-        return partial_function
+        """Return the doublet-DALI tensors (G, H) with shapes (P,P,P) and (P,P,P,P)."""
+        return self._lx.dali(n_workers=n_workers)
